# services/control-plane/internal/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velane Control Plane API",
    "version": "1.0.0",
    "description": "HTTP API for Velane authentication, tenant administration, workflows, integrations, invocation, observability, and embeds. Unless stated otherwise, authenticated operations accept either a Velane API key (`vl_...`), a session JWT, or the session cookie. The `x-required-scope` extension records the minimum API-key scope or session role."
  },
  "servers": [
    { "url": "/", "description": "Current Velane deployment" }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "sessionCookie": [] }
  ],
  "paths": {
...
    "/v1/tenant/members/invite": {
      "post": { "operationId": "inviteMember", "summary": "Invite a tenant member", "tags": ["Tenants"], "x-required-scope": "admin", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InviteMemberRequest" } } } }, "responses": { "201": { "description": "Invite created; the token is returned only once", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InviteMemberResponse" } } } }, "400": { "$ref": "#/components/responses/BadRequest" } } }
    },
    "/v1/tenant/members/{userID}": {
      "parameters": [{ "$ref": "#/components/parameters/UserID" }],
      "delete": { "operationId": "removeMember", "summary": "Remove a tenant member", "tags": ["Tenants"], "x-required-scope": "admin", "responses": { "204": { "$ref": "#/components/responses/NoContent" }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
...
    "/v1/invocations/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/InvocationID" }],
      "get": { "operationId": "getInvocation", "summary": "Get invocation details", "tags": ["Invocations"], "responses": { "200": { "description": "Invocation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invocation" } } } }, "404": { "$ref": "#/components/responses/NotFound" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Velane API key (`vl_...`) or RS256 session JWT." },
      "sessionBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "RS256 admin session JWT. API keys are not accepted by session-only endpoints." },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "velane_session", "description": "HttpOnly admin session cookie." },
      "embedToken": { "type": "http", "scheme": "bearer", "description": "Opaque embed token (`et_...`). Embed tokens never receive admin scope." },
      "gitWebhookSignature": { "type": "apiKey", "in": "header", "name": "X-Hub-Signature-256", "description": "HMAC-SHA256 signature in `sha256=<hex>` format." },
      "nangoWebhookSignature": { "type": "apiKey", "in": "header", "name": "X-Nango-Hmac-Sha256", "description": "Nango webhook HMAC signature." }
    },
    "parameters": {
...
      "InvocationID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "UserID": { "name": "userID", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
...
    "responses": {
      "NoContent": { "description": "Operation completed with no response body." },
      "BadRequest": { "description": "Invalid request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Forbidden": { "description": "The credential lacks access or the required scope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "The requested resource was not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } } },
...
      "Environment": { "type": "string", "enum": ["dev", "staging", "prod"] },
      "InvocationStatus": { "type": "string", "enum": ["pending", "running", "completed", "failed", "timeout", "oom_killed"] },
...
      "InviteMemberRequest": { "type": "object", "required": ["email"], "properties": { "email": { "type": "string", "format": "email" }, "role": { "type": "string", "enum": ["owner", "admin", "integration_manager", "manage", "member", "invoke", "viewer"], "default": "manage" } } },
...
      "InviteMemberResponse": { "type": "object", "required": ["invite_token", "expires_at"], "properties": { "invite_token": { "type": "string" }, "expires_at": { "type": "string", "format": "date-time" } } },
...
      "Invocation": { "type": "object", "required": ["id", "snippet_id", "version_id", "environment", "tenant_id", "status", "input_payload", "output", "stderr", "duration_ms", "peak_memory_mb", "cpu_ms", "created_at", "invoke_mode"], "properties": { "id": { "type": "string" }, "snippet_id": { "type": "string" }, "version_id": { "type": "string" }, "environment": { "$ref": "#/components/schemas/Environment" }, "tenant_id": { "type": "string" }, "status": { "$ref": "#/components/schemas/InvocationStatus" }, "input_payload": { "type": "string", "description": "Raw JSON encoded as a string." }, "input_ref": { "type": "string" }, "output": { "type": "string", "description": "Raw JSON encoded as a string." }, "output_ref": { "type": "string" }, "error": { "type": "string" }, "stderr": { "type": "string" }, "stderr_ref": { "type": "string" }, "duration_ms": { "type": "integer" }, "peak_memory_mb": { "type": "integer" }, "cpu_ms": { "type": "integer" }, "created_at": { "type": "string", "format": "date-time" }, "completed_at": { "type": ["string", "null"], "format": "date-time" }, "callback_url": { "type": "string", "format": "uri" }, "invoke_mode": { "type": "string", "enum": ["sync", "async", "stream"] }, "payload_state": { "type": "string", "enum": ["legacy", "pending", "stored", "failed", "purged"] } } },
...
    }
  }
}